pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-weakness-analyzer"
version = "0.1.0"
description = "Chess.com analytics toolkit for identifying player weaknesses."
authors = [{name = "Pierre Delattre"}]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["chess", "analytics", "stockfish", "streamlit", "cli"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "python-chess==1.999",
  "pandas==2.2.3",
  "pyarrow==16.1.0",
  "requests==2.31.0",
  "tqdm==4.66.1",
  "pydantic==2.4.2",
  "pydantic-settings==2.0.3",
  "streamlit==1.41.1",
  "matplotlib==3.9.2",
  "rich==13.6.0",
  "scipy==1.16.2",
  "typer[all]==0.9.0"
]

[project.optional-dependencies]
dev = [
  "pytest==7.4.2"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "N",
    "W"
]

[tool.ruff.format]
quote-style = "double"
